fix: added various fals states if the id is not found on teams, teams-members

+++ src/controller/team-members-controller.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import { z } from 'zod';
 import { prisma } from '../database/prisma';
+import { AppError } from '../utils/AppError';
 
 class TeamMembersController {
   async create(request: Request, response: Response) {
@@ -38,6 +39,10 @@ class TeamMembersController {
       include: { users: true },
     });
 
+    if (teamsMembers.length === 0) {
+      throw new AppError('There is no teams with this id');
+    }
+
     return response.status(201).json(teamsMembers);
   }
 
@@ -48,6 +53,19 @@ class TeamMembersController {
     });
 
     const { userId, teamsId } = bodySchema.parse(request.body);
+
+    const teamMembers = await prisma.teamMembers.findUnique({
+      where: {
+        userId_teamsId: {
+          userId,
+          teamsId,
+        },
+      },
+    });
+
+    if (!teamMembers) {
+      throw new AppError('This person in not on this team');
+    }
 
     await prisma.teamMembers.delete({
       where: {
